requirements.txt
Django
djangorestframework
psycopg2-binary
pymongo
elasticsearch
djongo
